[core] Add optional "defaultTime" argument
We are adding an optional "defaultTime" argument to the "getTimeParams",
which defaults to "last15Minutes". This allows to overwrite the default
selected time in for example dashboards.

This feature is now used for the cost management in the Azure plugin to
set the default selected time range to 30 days. The formaly selected 15
minutes time range doesn't returned any data from the Azure API.

We also fixed a small scrolling bug for the cost management page in the
Azure plugin.
